==> analysis/vl-modeling/02_analysis-vl-model-holte.Rmd <==
---
title: "Diagnositics and fits for the Holte model"
author: "Bryan Mayer"
date: "`r Sys.Date()`"
output: 
  html_document:
    keep_md: true
    code_folding: hide
---

# Overview

Here, we assess the RV217 Holte models to the AMP participants fit using various approaches in fit-vl-model-holte. 

The best approach was refitting the correlations in the placebo population ('refit_corr-placebo'). In the Final unadjusted Model section, further exploration is done for this model including simulating VL via an ODE model.

```{r load-packages, echo = T, message=F,output=F, warning=F}
knitr::opts_chunk$set(echo = T)
library(VISCfunctions)
library(tidyverse)
library(here)
library(glue)
library(kableExtra)
library(GGally)
source(here("R", "directory-funs.R"))
source(here("R", "mlx-model-funs.R"))
source(here("R", "mlx-res-processing.R"))
source(here("R", "mlx-plot-funs.R"))
vl_dir_tag = function(x) paste0("VL/", x)
vl_model_here = function(x, ...) mlx_model_here(vl_dir_tag(x), ...)

theme_set(theme_bw() + theme(legend.position = "top"))
strip_as_yaxis = theme(strip.placement = "outside", strip.background.y = element_blank())
options(scipen = 999)

rx_lab = tibble(
  rx_code2 = c("C", "T1", "T2"),
  trt_cat = c("Placebo", "VRC01 (10mg/kg)", "VRC01 (30mg/kg)")
)
rx_lab_pool = tibble(
  rx_code2 = c("C", "T1", "T2"),
  trt_cat_pool = c("Placebo", "VRC01 (Pooled)", "VRC01 (Pooled)")
)

save_shared_folder = F

```

```{r load-data}

time_data = read_csv(clean_data_here("adata-time-summary.csv"), col_types = cols()) %>%
  select(pub_id, contains("_day"))

demo_merge = read_csv(clean_data_here("adata-vl.csv"), col_types = cols()) %>%
  dplyr::select(pub_id, contains("gmt"), nisolates, protocol, study, rx_code2) %>% 
  distinct() %>%
  left_join(rx_lab, by = "rx_code2")  %>%
  left_join(time_data, by = "pub_id")

stopifnot(nrow(demo_merge) == 158)

dan_holte_est = read_csv(here("output/vl-analysis/estimatedIndividualParameters.txt"), col_types = cols()) %>%
  mutate(model = "Dan-Holte") %>%
  select(pub_id = id, model, contains("_mode"), -V0_mode, -aS_mode) %>%
  left_join(demo_merge, by = "pub_id")

dan_pop_parm =  read_csv(here("output/vl-analysis/populationParameters.txt"), col_types = cols())
dan_pop_ll = read_csv(here("output/vl-analysis/logLikelihood.txt"), col_types = cols()) %>%
                        rename(parameter = criteria, value = importanceSampling)
dan_pop_est = bind_rows(dan_pop_parm, dan_pop_ll) %>%
  mutate(model = "Dan")

paul_times = read_csv(clean_data_here("adata-time-summary.csv"), col_types = cols()) %>%
  filter(vl_model_flag) %>%
  mutate(model = "Poisson-fitter", tinf_est = fp_day - tpf_day) %>%
  select(pub_id, model, tinf_est, rx_code2)

```

```{r mlx-data-setup}

placebo_models = tribble(
  ~model_label, ~model_location,
  "RSIF-placebo", "holte_vl_corr_placebo",
  #"RSIF_nocorr-placebo", "holte_vl_placebo",
  "holte_refit_omega-placebo", "holte_vl_refitomega_placebo",
  "holte_refit_corr-placebo", "holte_vl_refitcorr_placebo",
  "holte_refit_omega_corr-placebo", "holte_vl_refitomegacorr_placebo", 
)

vrc01_models = tribble(
  ~model_label, ~model_location,
  "RSIF-vrc01", "holte_vl_corr_vrc01",
  #"RSIF_nocorr-vrc01", "holte_vl_vrc01",
  "holte_refit_omega-vrc01", "holte_vl_refitomega_vrc01",
  "holte_refit_corr-vrc01", "holte_vl_refitcorr_vrc01",
  "holte_refit_omega_corr-vrc01", "holte_vl_refitomegacorr_vrc01", 
)

holte_models = tribble(
  ~model_label, ~model_location,
  "RSIF", "holte_vl_corr_all",
  #"RSIF_nocorr", "holte_vl_all",
  "holte_refit_omega", "holte_vl_refitomega_all",
  "holte_refit_corr", "holte_vl_refitcorr_all",
  "holte_refit_omega_corr", "holte_vl_refitomegacorr_all"
)

parameter_labels = tribble(
  ~parm, ~parameter_label,
  "aS", "alphaS",
  "log10_dS", "log10(deltaS)",
  "dS", "deltaS",
  "lBt0", "log10(beta)",
  "dI", "kappa",
  "log10_dI", "log10(kappa)",
  "n", "h",
  "lp", "log10(pi)",
  "initT", "First pos.- Inf. day"
)
```


```{r load-mlx-model-data}

pop_ests = stack_model_ests(vl_dir_tag(holte_models$model_location),  
                            holte_models$model_label) 

placebo_pop_ests = stack_model_ests(vl_dir_tag(placebo_models$model_location),  
                            placebo_models$model_label) 

all_indiv_obs = get_indiv_preds(vl_dir_tag(holte_models$model_location[1])) %>%
  rename(pub_id = ID) %>%
  mutate(weeks = time/7)

indiv_vl_parms_placebo = stack_indiv_parms(vl_dir_tag(placebo_models$model_location),
                                           placebo_models$model_label) %>%
  select(pub_id = id, model, study_region_cat, contains("_mode"), -V0_mode, -aS_mode) %>%
  left_join(demo_merge, by = "pub_id") 

indiv_vl_parms_vrc01 = stack_indiv_parms(vl_dir_tag(vrc01_models$model_location), 
                                         vrc01_models$model_label) %>%
  select(pub_id = id, model, study_region_cat, contains("_mode"), -V0_mode, -aS_mode) %>%
  left_join(demo_merge, by = "pub_id") 

indiv_vl_parms_all = stack_indiv_parms(vl_dir_tag(holte_models$model_location), 
                                       holte_models$model_label) %>%
  select(pub_id = id, model, study_region_cat, contains("_mode"), -V0_mode, -aS_mode) %>%
  left_join(demo_merge, by = "pub_id") 

fits_placebo_train = stack_indiv_fits(vl_dir_tag(placebo_models$model_location),
                                 placebo_models$model_label) %>%
  bind_rows(stack_indiv_fits(vl_dir_tag(vrc01_models$model_location),
                                 vrc01_models$model_label)
           )%>%
  mutate(weeks = time / 7, 
         model_general = str_remove(str_split_fixed(model, "-", n = 2)[,1], "holte_")) %>%
  rename(pub_id = ID) %>%
  left_join(demo_merge, by = "pub_id")

fits_pool = stack_indiv_fits(vl_dir_tag(holte_models$model_location),
                                 holte_models$model_label) %>%
  mutate(weeks = time / 7, 
         model_general = str_remove(model, "holte_")) %>%
  rename(pub_id = ID) %>%
  left_join(demo_merge, by = "pub_id")


```

# Population Parameters

```{r placebo-pop-ests}

placebo_pop_ests %>%
  dplyr::mutate(
    parameter = if_else(parameter == "OFV", "LogLik", parameter),
    parameter = if_else(parameter == "initT_pop", "First pos.- Inf. day", parameter),
    model = str_remove(model, "holte_"),
    estimate_rse = if_else(!is.na(rse_sa),
                           stat_paste(value, abs(rse_sa), digits = 2),
                           stat_paste(value, digits = 2)) 
    ) %>%
  dplyr::filter(!str_detect(parameter, "_pop")) %>%
  dplyr::select(model, parameter, estimate_rse) %>% 
  pivot_wider(values_from = estimate_rse, names_from = model)  %>%
  kable(format = "html", booktabs = TRUE, linesep = "", escape = FALSE,
        caption = "Placebo-trained model population parameters (%RSE). RSEs are not estimated for fixed parameters or the error term.") %>%
  kable_styling(full_width = F)

```

```{r pooledpop-ests}

pop_ests %>%
  bind_rows(dan_pop_est) %>%
  dplyr::mutate(
    parameter = if_else(parameter == "OFV", "LogLik", parameter),
    parameter = if_else(parameter == "initT_pop", "First pos.- Inf. day", parameter),
    model = str_remove(model, "holte_"),
    estimate_rse = if_else(!is.na(rse_sa),
                           stat_paste(value, abs(rse_sa), digits = 2),
                           stat_paste(value, digits = 2)) 
    ) %>%
  dplyr::filter(!str_detect(parameter, "_pop")) %>%
  dplyr::select(model, parameter, estimate_rse) %>% 
  pivot_wider(values_from = estimate_rse, names_from = model)  %>%
  kable(format = "html", booktabs = TRUE, linesep = "", escape = FALSE,
        caption = "Pooled model population parameters (%RSE). RSEs are not estimated for fixed parameters or the error term.") %>%
  kable_styling(full_width = F)

```

# Evaluate infection times

1. fitted to placebo 
2. two stage for vrc01
3. fitted to all

Consistency between models among 1., 2 and 3
Consistence between models/fits between 1 and 2; and 2 and 3 among placebo or treated

```{r}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  mutate(model_data = "placebo_trained", model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  bind_rows(mutate(subset(indiv_vl_parms_all), model_data = "full_model")) %>%
  select(pub_id, tinf_est = initT_mode, model, model_data) %>%
  ggplot(aes(x = model_data, y = tinf_est)) +
  geom_boxplot() +
  ylab("est. inf day (prior to first positive)") +
  geom_path(aes(group = pub_id), alpha = 0.15) +
  facet_wrap( ~ model, nrow = 1)


```

## Consistency within placebo groups

```{r}

indiv_vl_parms_placebo %>%
  mutate(model_data = "placebo_only", model = str_remove(model, "-placebo")) %>%
  bind_rows(mutate(subset(indiv_vl_parms_all, rx_code2 == "C"), model_data = "full_model")) %>%
  select(pub_id, tinf_est = initT_mode, model, model_data) %>%
  ggplot(aes(x = model_data, y = tinf_est)) +
  geom_boxplot() +
  ylab("Placebo est. inf day (prior to first positive)") +
  geom_path(aes(group = pub_id), alpha = 0.15) +
  facet_wrap( ~ model, nrow = 1)

indiv_vl_parms_placebo %>% 
  bind_rows(subset(indiv_vl_parms_all, rx_code2 == "C")) %>%
  select(pub_id, tinf_est = initT_mode, model) %>%
  pivot_wider(values_from = tinf_est, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "First pos - inf time",
          ylab = "First pos - inf time")

```

## Consistency within vrc01 groups

```{r}

indiv_vl_parms_vrc01 %>%
  mutate(model_data = "placebo_trained", model = str_remove(model, "-vrc01")) %>%
  bind_rows(mutate(subset(indiv_vl_parms_all, rx_code2 != "C"), model_data = "full_trained")) %>%
  select(pub_id, tinf_est = initT_mode, model, model_data) %>%
  ggplot(aes(x = model_data, y = tinf_est)) +
  geom_boxplot() +
  ylab("VRC01 est. inf day (prior to first positive)") +
  geom_path(aes(group = pub_id), alpha = 0.15) +
  facet_wrap( ~ model, nrow = 1)

indiv_vl_parms_vrc01 %>% 
  bind_rows(subset(indiv_vl_parms_all, rx_code2 != "C")) %>%
  select(pub_id, tinf_est = initT_mode, model) %>%
  pivot_wider(values_from = tinf_est, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "First pos - inf time",
          ylab = "First pos - inf time")

```


## Correlation plots (placebo trained models)

### Spearman (Default)

```{r, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  bind_rows(dan_holte_est) %>%
  dplyr::select(pub_id, tinf_est = initT_mode, model) %>%
  bind_rows(select(paul_times, -rx_code2)) %>%
  pivot_wider(values_from = tinf_est, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "First pos - inf time",
          ylab = "First pos - inf time")

```

### Lin's CCC

```{r time-est-scaled, fig.height=9, fig.width=9, warning = F, message=F}
my_custom_cor <- function(data, mapping, sizeRange = c(1, 5), ...) {

  # get the x and y data to use the other code
 
  x <-  GGally::eval_data_col(data, mapping$x)
  y <-  GGally::eval_data_col(data, mapping$y)

  if(sd(x) == 0 | sd(y) == 0){
    tt = "NA"
  } else{
    ct <- epiR::epi.ccc(x, y)
    r <- unname(ct$rho.c$est)
    rt <- format(r, digits = 2)[1]
    tt <- as.character(rt)
  }
  #tt = "test"

  # plot the cor value
  p <- ggally_text(
   label = tt, 
   mapping = aes(),
   xP = 40, yP = 40, 
   size = 4,
   ...
  ) +

theme(panel.background=element_blank(),
        panel.grid.minor=element_blank(),
        panel.grid.major=element_blank()) 

 p
}

my_custom_smooth <- function(data, mapping, ...) {
  ggplot(data = data, mapping = mapping) +
    geom_point() + 
    scale_color_brewer(palette="Accent") +
    geom_abline(color = I("black")) +
    theme(panel.background=element_blank(),
        panel.grid.minor=element_blank(),
        panel.grid.major=element_line(color="gray"))
}


my_boxplot = function (data, mapping, ..., boxPlot = TRUE) {
    ggplot(data = data, aes(y = 50))  + geom_boxplot(mapping, ..., width = 15) +
    coord_flip()
}

suppressMessages({
indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  bind_rows(dan_holte_est) %>%
  dplyr::select(pub_id, tinf_est = initT_mode, model) %>%
  bind_rows(select(paul_times, -rx_code2)) %>%
  pivot_wider(values_from = tinf_est, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(upper = list(continuous = wrap(my_custom_cor)), 
          lower = list(continuous = my_custom_smooth),
          diag = list(continuous = my_boxplot),
          progress = F, xlab = "First pos - inf time",
           ylab = "First pos - inf time") +
  scale_x_continuous(limits = c(4.5, 93)) +
  scale_y_continuous(limits = c(4.5, 93))
})

```

# Correlation between models among VL parameters

## Log beta

```{r lbeta-corr, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  bind_rows(dan_holte_est) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  dplyr::select(pub_id, lBt0_mode, model) %>%
  pivot_wider(values_from = lBt0_mode, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "log10 beta",
          ylab = "log10 beta")

```

## Log burst size

```{r lpi-corr, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  bind_rows(dan_holte_est) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  dplyr::select(pub_id, lp_mode, model) %>%
  pivot_wider(values_from = lp_mode, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "log10 pi",
          ylab = "log10 pi")

```

## Log S death rate

Note: this was fit log-normal, and logged for this plot

```{r dS-corr, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  bind_rows(dan_holte_est) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  dplyr::select(pub_id, dS_mode, model) %>%
  mutate(dS_mode = log10(dS_mode)) %>%
  pivot_wider(values_from = dS_mode, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "log10 dS",
          ylab = "log10 dS")

```

## I death rate

```{r dI-corr, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  bind_rows(dan_holte_est) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  dplyr::select(pub_id, dI_mode, model) %>%
  pivot_wider(values_from = dI_mode, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "dI",
          ylab = "dI")

```

##  holte power term (n)

```{r n-corr, fig.height=9, fig.width=9}

indiv_vl_parms_placebo %>%
  bind_rows(indiv_vl_parms_vrc01) %>%
  bind_rows(dan_holte_est) %>%
  mutate(model = str_remove(str_remove(model, "-placebo"), "-vrc01")) %>%
  dplyr::select(pub_id, n_mode, model) %>%
  pivot_wider(values_from = n_mode, names_from = model) %>%
  dplyr::select(-pub_id) %>% 
  ggpairs(progress = F, xlab = "n",
          ylab = "n")

```

# Correlation within parameters

```{r parm-cor}

parm_corr_plot = map(unique(indiv_vl_parms_placebo$model), function(i){
  mod_title = str_remove(str_remove(str_remove(i, "-placebo"), "-vrc01"), "holte_")
  indiv_vl_parms_placebo %>%
    bind_rows(indiv_vl_parms_vrc01) %>%
    filter(model == i) %>%
    dplyr::select(contains("_mode")) %>%
    ggpairs(progress = F, title = mod_title)
})

```

```{r cor-pl-loop, fig.cap=unique(indiv_vl_parms_placebo$model), fig.height=9, fig.width=9}

for(i in 1:length(unique(indiv_vl_parms_placebo$model))){
  print(parm_corr_plot[[i]])
}

```

```{r dan, fig.height=9, fig.width=9}


dan_holte_est %>%
  dplyr::select(contains("_mode")) %>%
  ggpairs(progress = F, title = "Dan model")

```

# Generate fits

```{r pooled-fits}

indiv_plots = map(split(sort(unique(fits_pool$pub_id)),
                         ceiling(seq_along(unique(fits_pool$pub_id))/(5 * 2))), 
                  function(id_set) {
                    obs_dat = dplyr::filter(all_indiv_obs, pub_id %in% id_set)
                    fits_pool %>%
                      dplyr::filter(pub_id %in% id_set) %>%
                      ggplot(aes(x = weeks, y = log10(pmax(10, indivPredMode)))) +
                      geom_line(aes(colour = model_general)) +
                      geom_point(data = obs_dat,
                                 aes(y = log10(DV), shape = as.factor(censored != 0))) +
                      scale_shape_discrete(guide = "none") +
                      facet_wrap(~ pub_id, nrow = 5, ncol = 2) +
                      coord_cartesian(xlim = c(-5, NA)) +
                      labs(y = "log10 VL") +
                      theme(legend.position = 'top')
                  })

ggsave(here("output/vl-modeling/pool-fits.pdf"), 
       width = 8, height = 10,
       gridExtra::marrangeGrob(indiv_plots, nrow=1, ncol=1))

```

```{r trained-fits}

indiv_train_plots = map(split(sort(unique(fits_placebo_train$pub_id)),
                         ceiling(seq_along(unique(fits_placebo_train$pub_id))/(5 * 2))), 
                  function(id_set) {
                    obs_dat = dplyr::filter(all_indiv_obs, pub_id %in% id_set)
                    fits_placebo_train %>%
                      dplyr::filter(pub_id %in% id_set) %>%
                      ggplot(aes(x = weeks, y = log10(pmax(10, indivPredMode)))) +
                      geom_line(aes(colour = model_general)) +
                      geom_point(data = obs_dat,
                                 aes(y = log10(DV), shape = as.factor(censored != 0))) +
                      scale_shape_discrete(guide = "none") +
                      facet_wrap(~ pub_id, nrow = 5, ncol = 2) +
                      coord_cartesian(xlim = c(-5, NA)) +
                      labs(y = "log10 VL") +
                      theme(legend.position = 'top')
                  })

ggsave(here("output/vl-modeling/placebo-train-fits.pdf"), 
       width = 8, height = 10,
       gridExtra::marrangeGrob(indiv_train_plots, nrow=1, ncol=1))

```

# Final Unadjusted Model

```{r setup-final-model, class.source = "fold-show"}

source(here("R", "rxode-funs.R"))
source(here("R", "summarize_sim_holte_vl.R"))
source(here("models", "rxode-models", "rxode-models.R"))

unadj_placebo_pop_ests = get_model_ests("VL/holte_vl_refitcorr_placebo") 

unadj_placebo_vl_parms = get_indiv_parms("VL/holte_vl_refitcorr_placebo")
unadj_vrc01_vl_parms = get_indiv_parms("VL/holte_vl_refitcorr_vrc01")

unadj_vl_parms = bind_rows(unadj_placebo_vl_parms, unadj_vrc01_vl_parms) %>%
  select(pub_id = id, contains("_mode"), study_region_cat) %>%
  left_join(demo_merge, by = "pub_id") 

save_shared_folder = F

```

## Simulate VL

```{r rxode-sim}

unadj_vl_sims = map_df(1:nrow(unadj_vl_parms), function(i){
  parm_set = unadj_vl_parms[i, ]
  stopifnot(nrow(parm_set) == 1)
  run_vl_model(c(seq(0, 14, by = 0.1), 15:225), holte_model, 
               theta = prep_vl_model_parms(mode_parms = parm_set)) %>%
  mutate(pub_id = parm_set$pub_id, weeks = time/7)
})

final_setpt = map_df(1:nrow(unadj_vl_parms), function(i){
  parm_set = unadj_vl_parms[i, ]
  stopifnot(nrow(parm_set) == 1)
  run_vl_model(950:1000, holte_model, 
               theta = prep_vl_model_parms(mode_parms = parm_set)) %>%
    mutate(pub_id = parm_set$pub_id)
}) %>%
  group_by(pub_id) %>%
  summarize(final_setpt = mean(log10V), .groups = "drop")

```

```{r spot-check}
set.seed(15)
test_ids = sample(unadj_vl_parms$pub_id, 25, replace = F)

mlx_model_fits = get_indiv_fits("VL/holte_vl_refitcorr_placebo") %>%
  bind_rows(get_indiv_fits("VL/holte_vl_refitcorr_vrc01"))  %>%
  rename(pub_id = ID) %>%
  left_join(select(unadj_vl_parms, pub_id, initT_mode), by = "pub_id") %>%
  mutate(
    time_adj = time + initT_mode,
    weeks = time_adj / 7, 
    log10V = log10(pmax(0.01, indivPredMode)),
         source = "mlx-sims") %>% 
  filter(weeks >= 0)

unadj_vl_sims %>%
  mutate(source = "mlx-rxode") %>%
  bind_rows(mlx_model_fits) %>%
  dplyr::filter(pub_id %in% test_ids) %>%
  ggplot(aes(x = weeks, y = log10V)) +
  geom_line(aes(colour = source)) +
  scale_shape_discrete(guide = "none") +
  facet_wrap( ~ pub_id, nrow = 5, ncol = 5) +
  coord_cartesian(xlim = c(-5, NA)) +
  labs(y = "log10 VL") +
  theme(legend.position = 'top') +
  ggtitle("spot-check")

```

## Model Summary Measures

```{r}

vl_summary = summarize_sim_holte_vl(vl_sims = unadj_vl_sims, 
                                    vl_parms = unadj_vl_parms,
                                    setpt_end = 90, auc_end = 90, 
                                    max_uplope_time = c(3, 21)) %>%
  left_join(final_setpt, by = "pub_id")

stopifnot(all(!is.na(vl_summary)))
stopifnot(nrow(vl_summary) == 158)

```

```{r plot-endpoints, fig.height = 8, fig.width = 7.5}

key_measures = c("upslope_r0", "unadjusted_r0", "model_log10peak", "geo_auc_3mo", "log10_auc_3mo", "model_peak_day",
                 "model_setpt")
ic80gte1_colors = c(`IC80 < 1 mcg/mL` = "#008080", `IC80 >= 1 mcg/mL` = "coral")

(vl_summary %>%
  left_join(rx_lab_pool, by = "rx_code2") %>%
  mutate(
    ic80gte1 = factor(gmt80ls >= 1, levels = c(F, T), labels = c("IC80 < 1 mcg/mL", "IC80 >= 1 mcg/mL"))
    ) %>%
  select(pub_id, key_measures, ic80gte1, trt_cat_pool) %>%
  pivot_longer(names_to = "measure", values_to = "value", cols = key_measures) %>%
  ggplot(aes(x = trt_cat_pool, y = value, colour = ic80gte1)) +
  geom_boxplot() +
  geom_point(position = position_dodge(width = 0.75)) +
  facet_wrap(~measure, strip.position = "left", scales = "free_y") +
  ylab("") + xlab("") +
  scale_color_manual("", values = ic80gte1_colors) +
  strip_as_yaxis +
    theme(legend.direction = "vertical")) %>%
  lemon::reposition_legend(position =  'center',  panel = 'panel-3-2')

```

### Measure comparisons

```{r setpt-comp}

vl_summary %>%
  dplyr::select(contains("setpt")) %>%
  ggpairs(progress = F)

```

```{r auc-comp}

vl_summary %>%
  dplyr::select(contains("auc")) %>%
  ggpairs(progress = F)

```

```{r r0-comp}

vl_summary %>%
  dplyr::select(contains("r0")) %>%
  ggpairs(progress = F)

```

## Save final model

```{r save-output, class.source = "fold-show"}

write_csv(unadj_placebo_pop_ests, clean_data_here("unadjusted_model_vl_popparms.csv"))

write_csv(vl_summary, clean_data_here("unadjusted_model_vl_summary.csv"))

```

```{r save-shared, eval = save_shared_folder}

write_csv(tibble(), 
           paste0("/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/!VL MODEL UPDATED - ",
                  Sys.Date()))

write_csv(placebo_pop_ests, "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/unadjusted_model_vl_popparms.csv")

write_csv(vl_summary, "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/unadjusted_model_vl_summary.csv")

```
