==> analysis/vl-modeling/04_vrc01-pkpd-model.Rmd <==
---
title: "VRC01 PKPD Holte Model"
author: "Bryan Mayer"
date: "`r Sys.Date()`"
output: 
  html_document:
    keep_md: true
    code_folding: hide
---

# Overview

Here, we fit the PKPD model combining the Holte VL model with the 2-cmpt PK model.

$$ \beta^{pkpd} = \beta \frac{1}{1+[\rho \frac{C(t)}{IC50}]^{-h}}$$

The rho parameter is the "potency reduction" and the only parameter fit in this model. The IC50 and hill slope (h) come from the data.

It is possible that a profiling approach may be necessary where we only try to fit rho among sensitivity viruses. That is, we profile over a threshold $\rho_T$ where $ \beta^{pkpd} = \beta$ explicitly when $IC50 > \rho_T$.

In the PKPD model itself, each individual is fixed to their individual PK parameters from the PK analysis (ie, as regressors). The VL parameters are fixed at a population level. Will fit a model with and without the IC80-adjusted log-$\pi$ identified in the placebo-only analysis (see 01_placebo_pkpd_model).

```{r load-packages, echo = T, message=F,output=F, warning=F}
knitr::opts_chunk$set(echo = T)
library(VISCfunctions)
library(tidyverse)
library(here)
library(glue)
library(kableExtra)
library(GGally)
library(lixoftConnectors)
source(here("R", "directory-funs.R"))
source(here("R", "mlx-model-funs.R"))
source(here("R", "mlx-envir-res.R"))
source(here("R", "mlx-res-processing.R"))
source(here("R", "fix-parm-funs.R"))
source(here("R", "mlx-plot-funs.R"))
options(knitr.kable.NA = '')

pkpd_model_tag = function(x) paste0("PKPD/", x)

rx_lab = tibble(
  rx_code2 = c("C", "T1", "T2"),
  trt_cat = c("Placebo", "VRC01 (10mg/kg)", "VRC01 (30mg/kg)")
)

theme_set(theme_bw() + theme(legend.position = "top"))
strip_as_yaxis = theme(strip.placement = "outside", strip.background.y = element_blank())

initializeLixoftConnectors(software = "monolix")
options(scipen = 999)

```

```{r load-funs-data}

time_data = read_csv(clean_data_here("adata-time-summary.csv"), col_types = cols()) %>%
  select(pub_id, contains("_day"))

demo_merge = read_csv(clean_data_here("adata-vl.csv"), col_types = cols()) %>%
  dplyr::select(pub_id, contains("gmt"), nisolates, protocol, study, rx_code2) %>% 
  distinct() %>%
  left_join(rx_lab, by = "rx_code2")%>%
  left_join(time_data, by = "pub_id")
stopifnot(nrow(demo_merge) == 158)

raw_data = read_csv(mlx_data_here("amp-mlx-pkpd-holteT0.csv"), col_types = cols())

unadj_placebo_pop = read_csv(clean_data_here("unadjusted_model_vl_popparms.csv"), 
                             col_types = cols())
adj_placebo_pop = read_csv(clean_data_here("adjusted_placebo_vl_popparms.csv"), 
                             col_types = cols())

unadj_placebo_init = unadj_placebo_pop %>%
  select(parameter, value) %>% spread(parameter, value)

adj_placebo_init = adj_placebo_pop %>%
  select(parameter, value) %>% spread(parameter, value)

```

# Model fitting

```{r vl-dose-setup}
# check_data("amp-mlx-pkpd-holteT0.csv")
# check_data("amp-mlx-pkpd-indirect-holteT0.csv")

pkpd_types = tribble(
    ~var_name, ~type,
    "pub_id", "id",
    "TIME", "TIME",
    "DV", "observation",
    "cens", "cens",
    "AMT", "amount",
    "ADM", "admid",
    "RATE", "rate",
    "limit", "limit",
    "rx_code2", "catcov",
    "protocol", "catcov",
    "study_region_cat", "catcov",
    "gmt50ms", "ignore",
    "gmt80ms", "ignore",
    "gmt50ls", "regressor",
    "gmt80ls", "contcov",
    "gmt50mf", "ignore",
    "gmt80mf", "ignore",
    "hill_ls", "regressor",
    "hill_ms", "ignore",
    "hill_mf", "ignore",
    "Cl", "regressor",
    "V1", "regressor",
    "Q", "regressor",
    "V2", "regressor"
  ) 

stopifnot(ncol(raw_data) == nrow(pkpd_types))
stopifnot(all(pkpd_types$var_name == names(raw_data)))

```

```{r adj-model-setup, eval = F}


#these functions are run everytime

setup_pkpd_model = function(inits, mlx_data = "amp-mlx-pkpd-holteT0.csv"){
  
  amp_recipe(
    model_text_file = "PKPD_Holte.txt",
    data_file = mlx_data,
    variable_types = pkpd_types,
    initial_estimates = TRUE,
    obs_dist = "lognormal",
    error_model = "constant"  
    )
  
  setPopulationParameterInformation(a = list(initialValue = inits$a, 
                                              method = "FIXED"))
  set_holte_parms(holte_parms = inits, 
                  est_omega = "FIXED", 
                  est_corr_rsif = "FIXED",
                  dose = T)
  setIndividualParameterDistribution(lrho = "normal")
  setPopulationParameterInformation(lrho_pop = list(initialValue = 0))
  # from mlx documentation: https://www.youtube.com/watch?v=H4WVkTXc45I
  addContinuousTransformedCovariate(log_ic80_ctr = "log10(gmt80ls) -  0.638")
  
  check_init_assignment(inits, exclude_check = "")
  
}

```

# Models

## Adjusted PKPD

Using IC80 adjustment on log pi. Uses infection times from indirect model applied to VRC01 group.

```{r adj-pkpd-modeling, eval = F}

setup_pkpd_model(adj_placebo_init, mlx_data = "amp-mlx-pkpd-indirect-holteT0.csv")

setCovariateModel(lp = c(log_ic80_ctr = TRUE))
setPopulationParameterInformation(
  beta_lp_log_ic80_ctr = list(initialValue = adj_placebo_init$beta_lp_log_ic80_ctr, 
                              method = "FIXED")
)
getPopulationParameterInformation()

initial_runs()

save_project(pkpd_model_tag("adj-pkpd-model"))


# -- no omega for log rho

setup_pkpd_model(adj_placebo_init, mlx_data = "amp-mlx-pkpd-indirect-holteT0.csv")

setCovariateModel(lp = c(log_ic80_ctr = TRUE))
setPopulationParameterInformation(
  beta_lp_log_ic80_ctr = list(initialValue = adj_placebo_init$beta_lp_log_ic80_ctr, 
                              method = "FIXED")
)
setIndividualParameterVariability(lrho = F)
getPopulationParameterInformation()

initial_runs()

save_project(pkpd_model_tag("adj-pkpd-model-no-omega"))
save_project(pkpd_model_tag("adj-pkpd-model-no-omega-CI"))

setPopulationParameterInformation(lrho_pop = list(initialValue = 2.7982272104, method = "FIXED"))

```


```{r adj-pkpd-modeling-rho-1, eval = F}

setup_pkpd_model(adj_placebo_init, mlx_data = "amp-mlx-pkpd-indirect-holteT0.csv")

setCovariateModel(lp = c(log_ic80_ctr = TRUE))
setPopulationParameterInformation(
  beta_lp_log_ic80_ctr = list(initialValue = adj_placebo_init$beta_lp_log_ic80_ctr, 
                              method = "FIXED")
)

setPopulationParameterInformation(lrho_pop = list(initialValue = 0, method = "FIXED"))
setIndividualParameterVariability(lrho = F)
getPopulationParameterInformation()

set_project_settings(initial_estimates = FALSE)
runScenario()
get_pop_parms()

save_project(pkpd_model_tag("adj-pkpd-model-rhoEQ1"))

```


```{r CIs, eval = F}

library(Rsmlx)
ci = confintmlx(mlx_model_here(pkpd_model_tag("adj-pkpd-model-no-omega-CI"), mlxtran = T), 
           method="fim", parameters = c("lrho"), linearization = F, level = 0.95)

# this would take about two hours to run 25, so a work day to run 100
# confintmlx(mlx_model_here(pkpd_model_tag("adj-pkpd-model-no-omega-CI"), mlxtran = T), 
#            method="bootstrap", parameters = c("lrho"), linearization = F, nboot = 1, level = 0.95)

```

```{r profile-adj, eval = F}

lrho_range = c(1, log10(50), seq(2, 3, by = 0.1), 3.25, 3.5, 4, 4.5, 5)

adj_models_loglik = map_df(lrho_range, function(r){
  
  seed = round(r * 10)
  set.seed(seed)
  setup_pkpd_model(adj_placebo_init, mlx_data = "amp-mlx-pkpd-indirect-holteT0.csv")

  setCovariateModel(lp = c(log_ic80_ctr = TRUE))
  setIndividualParameterVariability(lrho = F)

  setPopulationParameterInformation(
    beta_lp_log_ic80_ctr = list(initialValue = adj_placebo_init$beta_lp_log_ic80_ctr, 
                              method = "FIXED"),
    lrho_pop = list(initialValue = r, method = "FIXED")
  )

  set_project_settings(initial_estimates = FALSE)
  runScenario()
  # sc = getScenario()
  # sc$tasks = c(populationParameterEstimation = T, conditionalModeEstimation = F, conditionalDistributionSampling = T,
  #              standardErrorEstimation = F, logLikelihoodEstimation = T, plots = F)
  # setScenario(sc)
  # 
  
  ll = getEstimatedLogLikelihood() %>% flatten_df() %>% rename(loglik = OFV)
  
  out = bind_cols(tibble(lrho = r, ll)) %>%
    mutate(seed = seed)
  print(out)
  out
})


adj_models_loglik %>%
  ggplot(aes(x = 10^lrho, y = loglik)) +
  geom_point() +
  geom_line() +
  scale_x_log10() +
  geom_hline(, colour = "red", yintercept = 9261 + 0.5 *qchisq(0.95, 1)) +
  annotate("point", colour = "red", y = 9261, x = 10^2.8, shape = 13
           ) +
  coord_cartesian(ylim = c(9250, 9700))

adj_models_loglik2 = adj_models_loglik %>%
  mutate(min_loglik = min(loglik))

adj_models_loglik2 %>%
  ggplot(aes(x = 10^lrho, y = pmin(abs(min_loglik-loglik), 20))) +
  geom_point() +
  geom_line(data = filter(adj_models_loglik2, lrho <= 2.2)) +
  geom_smooth(data = filter(adj_models_loglik2, lrho >= 2.2), se = F,  method= "loess") +
  scale_x_log10()

cowplot::plot_grid(x1, x2, nrow = 1)

write_csv(adj_models_loglik, "rho_profile.csv")


```

## Undjusted PKPD

```{r unadj-pkpd-models, eval = F}

setup_pkpd_model(adj_placebo_init)
getPopulationParameterInformation()

initial_runs()

save_project(pkpd_model_tag("pkpd-model"))


# -- no omega for log rho

setup_pkpd_model(adj_placebo_init)
setIndividualParameterVariability(lrho = F)
getPopulationParameterInformation()

initial_runs()

save_project(pkpd_model_tag("pkpd-model-no-omega"))

```

# Results

```{r results-setup}

pkpd_models = tribble(
  ~model_label, ~model_location,
  "no effect", "VL/holte_vl_refitcorr_vrc01",
  "indirect only", "VL-adjusted-placebo/cns-pi-vrc01",
  "pkpd only", "PKPD/pkpd-model",
  "pkpd only (no rho omega)", "PKPD/pkpd-model-no-omega",
  "indirect + pkpd", "PKPD/adj-pkpd-model",
  "indirect + pkpd (no rho omega)", "PKPD/adj-pkpd-model-no-omega",
  "indirect + pkpd (rho = 1)", "PKPD/adj-pkpd-model-rhoEQ1")


pkpd_pop_ests = stack_model_ests(pkpd_models$model_location,  
                                 pkpd_models$model_label) 

pkpd_fits = stack_indiv_fits(
  pkpd_models$model_location,
  pkpd_models$model_label
) %>%
  mutate(weeks = time / 7) %>%
  rename(pub_id = ID) %>%
  left_join(demo_merge, by = "pub_id") %>%
  mutate(
    weeks = if_else(model %in% c("no effect", "indirect only"),
                    weeks + fp_day/7,
                    weeks)
  )

all_indiv_obs = get_indiv_preds(pkpd_models$model_location[3]) %>%
  rename(pub_id = ID) %>%
  mutate(weeks = time/7) %>%
  left_join(demo_merge, by = "pub_id")

all_indiv_parms = stack_indiv_parms(pkpd_models$model_location,
  pkpd_models$model_label) %>%
  rename(pub_id = id) %>%
  left_join(demo_merge, by = "pub_id")

```

```{r model-criteria}

# this comes from 05_, so a bit circular
# for use in polishing the parameter plots

pkpd_model_paper_labels = tribble(
  ~model_location, ~model_label_paper, ~paper_model,
  "VL/holte_vl_refitcorr_vrc01", "VRC01_PKPD_Unadjusted (RV217 refit correlation)", F,
  "VL-adjusted-placebo/cns-pi-vrc01", "VRC01_PKPD_Indirect only (IC80-adjusted viral burst)", F,
  "PKPD/pkpd-model", "VRC01_PKPD_PKPD VRC01-adjusted only (w/ potency reduction variance)", F,
  "PKPD/pkpd-model-no-omega", "VRC01_PKPD_PKPD VRC01-adjusted only", F,
  "PKPD/adj-pkpd-model-rhoEQ1", "VRC01_PKPD_Indirect + PKPD VRC01-adjusted (w/o potency reduction (rho = 1))", F,
  "PKPD/adj-pkpd-model", "VRC01_PKPD_Indirect + PKPD VRC01-adjusted (w/ potency reduction variance)", F,
  "PKPD/adj-pkpd-model-no-omega", "VRC01_PKPD_Indirect + PKPD VRC01-adjusted", T
) %>%
  left_join(pkpd_models, by = "model_location")
  
paper_model_criteria = pkpd_pop_ests %>%
  left_join(select(pkpd_model_paper_labels, model_label, model_label_paper),
            by = c("model" = "model_label")) %>%
  mutate(
    parameter = if_else(parameter == "OFV", "LogLik", parameter)
  ) %>%
  filter(parameter == "LogLik" | parameter == "BICc") %>% 
  select(-se_sa, -rse_sa) %>%
  spread(parameter, value)


```

## Pop parms

```{r pop-table}
options(knitr.kable.NA = '')
pkpd_pop_ests %>%
  dplyr::mutate(
    parameter = if_else(parameter == "OFV", "LogLik", parameter),
    parameter = if_else(parameter == "lrho_pop", "log10 rho", parameter),
    parameter = if_else(parameter == "omega_lrho", "omega log10 rho", parameter)
    ) %>%
  dplyr::filter(!str_detect(parameter, "_pop") & 
                  !str_detect(parameter, "omega_") &
                   !str_detect(parameter, "corr_") &
                  parameter != "beta_lp_log_ic80_ctr"&
                  parameter != "a") %>%
  dplyr::select(model, parameter, value) %>% 
  pivot_wider(values_from = value, names_from = model) %>%
  mutate(parameter = 
           forcats::fct_relevel(factor(parameter), c("log10 rho", "omega log10 rho"))) %>%
  arrange(parameter) %>%
  kable(format = "html", booktabs = TRUE, linesep = "", escape = FALSE, digits = 2,
        caption = "PKPD models of VRC01 cohort. Indirect = IC80-adjusted burst size..") %>%
  kable_styling(full_width = F)


```

## MLX Fits 

```{r pkpd-fits}

indiv_plots = map(split(sort(unique(pkpd_fits$pub_id)),
                         ceiling(seq_along(unique(pkpd_fits$pub_id))/(5 * 2))), 
                  function(id_set) {
                    obs_dat = dplyr::filter(all_indiv_obs, pub_id %in% id_set)
                    pkpd_fits %>%
                      dplyr::filter(pub_id %in% id_set) %>%
                      ggplot(aes(x = weeks - fp_day/7, y = log10(pmax(10, indivPredMode)))) +
                      geom_line(aes(colour = model)) +
                      geom_point(data = obs_dat,
                                 aes(y = log10(DV), shape = as.factor(censored != 0))) +
                      scale_shape_discrete(guide = "none") +
                      facet_wrap(~ pub_id, nrow = 5, ncol = 2) +
                      coord_cartesian(xlim = c(-5, NA)) +
                      labs(y = "log10 VL") +
                      theme(legend.position = 'top')
                  })

ggsave(here("output/pkpd-modeling/pkpd-fits.pdf"), 
       width = 8, height = 10,
       gridExtra::marrangeGrob(indiv_plots, nrow=1, ncol=1))

```

## Indiv parms

```{r vrc01-parms, dev = "pdf", fig.width=8, fig.height=7, warning = F}
#   
#     `Virus-adjusted` = factor(`Virus-adjusted`,
#                              levels = c("Fixed RV217", "Unadjusted", "Adjusted", "PKPD"))


pkpd_parameters_labels <- tribble(
  ~parameter, ~parameter_parsed,
  "dS_mode", "log[10]~delta[S]",
  "lBt0_mode", "log[10]~beta",
  "lp_mode", "log[10]~pi",
  "dI_mode", "kappa",
  "n_mode", "n",
  "lrho_mode", "log[10]~rho"
)

all_indiv_parms %>%
  select(pub_id, model, contains("_mode"), -initT_mode, -V0_mode, -aS_mode) %>%
  gather("parameter", "value", -pub_id, -model) %>%
  left_join(paper_model_criteria, by = c("model")) %>%
  mutate(
     value =  if_else(parameter == "dS_mode", log10(value), value), # warning from other parameters in F
    parameter_plot = factor(parameter, levels = pkpd_parameters_labels$parameter,
                                 labels = pkpd_parameters_labels$parameter_parsed)) %>%
  separate(model_label_paper, into = c("Cohort", "virus_adj", "model_desc"), sep = "_", remove = F) %>%
  ggplot(aes(x = reorder(model_desc, BICc, min), y = value)) +
  geom_boxplot() +
  facet_wrap(~parameter_plot, 
             labeller = label_parsed,
             scales ="free_x", strip.position = "bottom") +
  theme(panel.grid = element_blank(), 
        strip.placement = "outside", 
        strip.background = element_blank()) +
  labs(x = "", y= "") +
  coord_flip()


```

```{r vrc01-parms-beta-only, dev = "pdf"}


all_indiv_parms %>%
  select(pub_id, model, contains("_mode"), -initT_mode, -V0_mode, -aS_mode) %>%
  gather("parameter", "value", -pub_id, -model) %>%
  left_join(paper_model_criteria, by = c("model")) %>%
  mutate(parameter_plot = factor(parameter, levels = pkpd_parameters_labels$parameter,
                                 labels = pkpd_parameters_labels$parameter_parsed)) %>%
  separate(model_label_paper, into = c("Cohort", "virus_adj", "model_desc"), sep = "_", remove = F) %>%
  filter(parameter == "lBt0_mode") %>%
  ggplot(aes(x = reorder(model_desc, BICc, min), y = value)) +
  geom_boxplot() +
  facet_wrap(~parameter_plot, 
             labeller = label_parsed,
             scales ="free_x", strip.position = "bottom") +
  theme(panel.grid = element_blank(), 
        strip.placement = "outside", 
        strip.background = element_blank()) +
  labs(x = "", y= "") +
  coord_flip()

```


```{r parms-sens}

all_indiv_parms %>%
  filter(gmt80ls < 1) %>%
  select(pub_id, model, contains("_mode"), -initT_mode, -V0_mode, -aS_mode) %>%
  gather("parameter", "value", -pub_id, -model) %>%
  ggplot(aes(x = model, y = value)) +
  geom_boxplot() +
  facet_wrap(~parameter, scales ="free_y") +
  ggtitle("sensitive viruses (ic80 < 1)")  +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

```

## Rho Comparisons

Based on the placebo model, sensitive viruses have higher initial VL. The PKPD model would have an opposing effect because sensitive viruses are neutralized when VRC01 is present. At the population average burst size (fairly resistant virus) the burst size would be under-estimated for sensitive viruses and the model find an optima with a high potency reduction (the PKPD effect doesnt seem to be doing much). 

Without omega, the estimates are similar and lower. Why this is dependent on the omega specification is complex and seemingly related to the flexibility of model draws and complexity of parameter correlations. Across all participants, the variation is fairly hidden as most viruses are resistant. However, among participants with sensitive viruses (< 1), there is variability of the parameter draws and changes in the medians across models that is shown above. 


```{r rho-setup}

pkpd_models_rho_omega = tribble(
  ~model_label, ~model_location,
  "pkpd only", "PKPD/pkpd-model",
  "indirect + pkpd", "PKPD/adj-pkpd-model",
)

pkpd_models_pkpd_only = tribble(
  ~model_label, ~model_location,
  "pkpd only (no rho omega)", "PKPD/pkpd-model-no-omega"
)

pkpd_models_final = tribble(
  ~model_label, ~model_location,
  "indirect + pkpd (no rho omega)", "PKPD/adj-pkpd-model-no-omega",
)

rho_indiv_parms = stack_indiv_parms(pkpd_models_rho_omega$model_location,
                                pkpd_models_rho_omega$model_label)


```

```{r}

vrc01_pop_ests_wide = pkpd_pop_ests %>%
  select(model, parameter, value) %>%
  spread(parameter, value)

ggplot(rho_indiv_parms, aes(x = model, y = 10^lrho_mode)) +
  geom_boxplot() +
  geom_point() +
  scale_y_log10(breaks = c(500, 1000, 1500)) +
  geom_point(data = vrc01_pop_ests_wide, aes(y = 10^lrho_pop),
             shape = "x", size = 4, colour = "red")

```
