==> analysis/vl-modeling/06_amp-sim-analysis.Rmd <==
---
title: "Simulating PKPD using AMP settings"
author: "Bryan Mayer"
date: "`r Sys.Date()`"
output: 
  html_document:
    keep_md: true
    code_folding: hide
---

# Overview

This is to generate the final figures from the PKPD model describing the population results in
the AMP study.

- Placebo effect plot (the burst size adjustment model)
- PKPD plots showing direct and indirect effects
- Main figure showing re-simulations with different VRC01 potency


```{r load-packages, echo = T, message=F,output=F,warning =F}
knitr::opts_chunk$set(echo = T)
library(here)
library(glue)
library(kableExtra)
library(RxODE)
library(viridis)
library(cowplot)
source(here("R", "directory-funs.R"))
source(here("R", "mlx-res-processing.R"))
source(here("R", "rxode-funs.R"))
source(here("models", "rxode-models", "rxode-models.R"))
source(here("models", "rxode-models", "rxode-pkpd-models.R"))
library(tidyverse)

save_shared_folder = F
rerun_sims = F

# added the placebo sims later
save_placebo = F
save_shared_folder_placebo = T & save_placebo

theme_set(theme_bw() + theme(legend.position = "top"))
strip_as_yaxis = theme(strip.placement = "outside", strip.background.y = element_blank())

titer_neut = function(id50, h = 1) 1 - beta_titer_neut(id50, h)
beta_titer_neut = function(id50, h = 1) 1/(1+id50^h)

final_vl_summary = read_csv(clean_data_here("final_vl_summary.csv"), col_types = cols()) %>%
  mutate(ic80_cat = if_else(gmt80ls < 1, "IC80 < 1 mcg/mL", "IC80 >= 1 mcg/mL"))

vrc01_dose =  read_csv(mlx_data_here("amp-mlx-pkpd-indirect-holteT0.csv"), col_types = cols()) %>%
  filter(ADM == 1) %>%
  distinct(pub_id, TIME, AMT, RATE) %>%
  mutate(AMT = as.numeric(AMT), RATE = as.numeric(RATE)) %>%
  arrange(pub_id, TIME) %>%
  group_by(pub_id) %>%
  mutate(DOSENO = row_number()) %>%
  ungroup()

ic80gte1_colors = c(`IC80 < 1 mcg/mL` = "#1F77B4", `IC80 >= 1 mcg/mL` = "coral")
ic80_label = c(`IC80 < 1 mcg/mL` =  bquote("#1F77B4"),
               `IC80 >= 1 mcg/mL` =  "coral"
             )

sim_arrange = tibble(
           levels = c("VRC01", "NoTreatment", "VRC01x10", "VRC01x100", "VRC01x1000"),
           labels = c("VRC01 (AMP design)", "No VRC01", 
                      "10x potency", "100x potency", "1000x potency")
         )
                   

# like the arms are: czl=['gray','lightblue','tab:green']
# and the 4 groups are czl2=['navy','tab:red','tab:blue','coral']

```

```{r amp-pe-data}

pt80_efficacy = tibble(
  PT80 = c(32, 82, 194),
  PE = c(50, 75, 90),
  PE_pct = paste0("PE ", PE, "%"),
  pe_neut = PE/100
)

empirical_amp_fit = drc::drm(pe_neut ~ PT80, data = pt80_efficacy, 
                        fct = drc::LL.5(names = c("slope", "lower", "upper",
                                                  "inflection", "asymmetry"),
                             fixed = c(NA, 0, 1, NA, NA)))

```


```{r load-placebo-mod}
raw_placebo_data = read_csv(mlx_data_here("amp-mlx-vl-placebo.csv"), col_types = cols())

placebo_ic80_range = c(min(raw_placebo_data$gmt80ls), max(raw_placebo_data$gmt80ls))
placebo_ic50_range = c(min(raw_placebo_data$gmt50ls), max(raw_placebo_data$gmt50ls))

final_pop_ests = read_csv(clean_data_here("final_model_popparms.csv"), col_types = cols())
final_pop_ests_wide = final_pop_ests %>%
  dplyr::select(parameter, value) %>% spread(parameter, value)

theta_vl_placebo = prep_vl_model_parms(mode_parms = final_pop_ests_wide, parm_tag = "_pop")

# takes IC80 and pop parms, returns adjusted burst size
burst_effect = function(ic80, pop_parms = final_pop_ests_wide) {
  lp = pop_parms$lp_pop + pop_parms$beta_lp_log_ic80_ctr * (log10(ic80) - 0.638)
  10^(lp)
}

# takes adjusted individual burst size, ic80, beta, returns unadjusted burst size
null_lp = function(lp_adj, ic80, pop_parms = final_pop_ests_wide) {
  lp_adj - pop_parms$beta_lp_log_ic80_ctr * (log10(ic80) - 0.638)
}

```

```{r load-trt-mod}

raw_vrc01_data = read_csv(mlx_data_here("amp-mlx-vl-vrc01.csv"), col_types = cols())

vrc01_ic80_range = c(min(raw_vrc01_data$gmt80ls), max(raw_vrc01_data$gmt80ls))
vrc01_ic50_range = c(min(raw_vrc01_data$gmt50ls), max(raw_vrc01_data$gmt50ls))

vrc01_pkpd_data = read_csv(here('output/pk-modeling/sim-pk-withVLneut.csv'), col_types = cols())
vrc01_first_pos = subset(vrc01_pkpd_data, first_positive) %>%
  mutate(PT80 = centr_conc/gmt80ls)

theta_vl_vrc01 = theta_vl_placebo

```

```{r pd-setup}

vl_summary = read_csv(clean_data_here("final_vl_summary.csv"), col_types = cols())

#1.28, use 1.25
mean_hill = mean(vl_summary$hill_ls)

amp_theta_pd = tibble(IC80 = NA, rho = 10^final_pop_ests_wide$lrho_pop, h = 1.25)

kable(amp_theta_pd, caption = "AMP PD parameters") %>% kable_styling(full_width = F)

```

```{r pk-setup}
pk_parms = tribble(
  ~parameter, ~value,
  "Cl", 0.575,
  "V1", 3.532,
  "V2", 4.749,
  "Q", 0.676
)

kable(pk_parms, caption = "AMP PK parameters") %>% kable_styling(full_width = F)

theta_pk = create_theta(pk_parms)

conc_range = range(vrc01_pkpd_data$centr_conc, na.rm = T)

```

# Individual trajectory simulations

## Placebo

Re-simulate placebo with and without indirect effect.

```{r sim-indiv-placebo}

vl_sims_placebo = map_df(subset(final_vl_summary, rx_code2 == "C" & !is.na(initT_mode))$pub_id, function(i){
  parm_set = subset(final_vl_summary, pub_id == i)
  stopifnot(nrow(parm_set) == 1)
  
  fp_day = parm_set$fp_day
  infection_day = parm_set$model_infection_day

  dosing = filter(vrc01_dose, pub_id == parm_set$pub_id)

  mtime = c(0:(fp_day + 80))

  theta_pk = select(parm_set, Cl, V1, Q, V2)
  theta_pd = tibble(IC80 = parm_set$gmt50ls * 4 ^(1/parm_set$hill_ls), 
                    rho = 10^parm_set$lrho_mode, 
                    h = parm_set$hill_ls)
  
  parm_set$lp_adj = parm_set$lp_mode
  parm_set$lp_unadj = null_lp(parm_set$lp_mode, ic80 = parm_set$gmt80ls)
  
  prep_vl_parms_burst = function(parm_set, adj_lp = T){
    if(!adj_lp) parm_set$lp_mode = parm_set$lp_unadj
    prep_vl_model_parms(mode_parms = parm_set)
  }  
    
  map_df(c(T, F), 
         ~(run_pkpd_models(mtime,      
                  holte_pkpd_model,
                  theta_pk = theta_pk,
                  theta_vl = prep_vl_parms_burst(parm_set, adj_lp = ..1),
                  theta_pd = theta_pd,
                  infection_time = infection_day) %>%
             mutate(indirect = ..1))) %>%
    filter(time > 0) %>%
    bind_cols(parm_set) %>%
    mutate(first_positive_row = time == fp_day) %>%
    select(pub_id, days = time, everything())
})  %>%
  mutate(days_fp = days - fp_day) %>%
  filter(days_fp > -30 & days >= model_infection_day) %>%
  rename(days_enrollment = days)

```

## VRC01 

Re-simulate VRC01 participant VLs with no VRC01, original dosing, 10x dosing, and 100x dosing.

```{r sim-indiv-vrc01, eval = rerun_sims}

vl_sims_vrc01 = map_df(subset(final_vl_summary, rx_code2 != "C" & !is.na(Cl))$pub_id, function(i){
  parm_set = subset(final_vl_summary, pub_id == i)
  stopifnot(nrow(parm_set) == 1)
  
  fp_day = parm_set$fp_day
  infection_day = parm_set$model_infection_day

  dosing = filter(vrc01_dose, pub_id == parm_set$pub_id)

  mtime = c(0:(fp_day + 80))

  theta_pk = select(parm_set, Cl, V1, Q, V2)
  theta_pd = tibble(IC80 = parm_set$gmt50ls * 4 ^(1/parm_set$hill_ls), 
                    rho = 10^parm_set$lrho_mode, 
                    h = parm_set$hill_ls)
  
  dose_mult = function(dose_data, multiplier = 1) mutate(dose_data, AMT = AMT * multiplier)
  
  run_mod = function(dose_in){
      run_pkpd_models(mtime,      
                  holte_pkpd_model,
                  theta_pk = theta_pk,
                  theta_vl = prep_vl_model_parms(mode_parms = parm_set),
                  theta_pd = theta_pd,
                  infection_time = infection_day,
                  infusion_pk_dosing = dose_in) %>% 
      filter(time > 0)
  }

  # both functions here are declared in this map
  map2_df(list(NULL, dosing, dose_mult(dosing, 10), dose_mult(dosing, 100), dose_mult(dosing, 1000)),
          c("NoTreatment", "VRC01", "VRC01x10", "VRC01x100", "VRC01x1000"),
          ~mutate(run_mod(..1), sim_type = ..2)) %>%
    bind_cols(parm_set) %>%
    mutate(first_positive_row = time == fp_day) %>%
    select(pub_id, days = time, everything())
})  %>%
  mutate(days_fp = days - fp_day) %>%
  filter(days_fp > -30 & days >= model_infection_day) %>%
  rename(days_enrollment = days)

```

```{r load-sims, eval = !rerun_sims}

vl_sims_vrc01 = read_csv(clean_data_here("amp-individual-vrc01-sims.csv"), show_col_types = F) %>%
  mutate(
    ic80_cat2 = factor(
      ic80_cat,
      levels = c("IC80 < 1 mcg/mL", "IC80 >= 1 mcg/mL"),
      labels = c("VRC01 sensitive: IC80<1",
                 "VRC01 resistant: IC80\u22651")
    ),
    sim_type2 = factor(sim_type, levels = sim_arrange$levels, labels =
                         sim_arrange$labels)
  )

```

# Placebo Sensitivity effects

the effect on log10 p:

$$lp = lp_0 + \beta_{IC80} (log_{10}IC80 - 0.638)$$

where 0.638 is the mean of the log10 IC80.

```{r}

# overlay the individual draws, at peak I

tibble(
  ic80 = seq(min(raw_placebo_data$gmt80ls), max(raw_placebo_data$gmt80ls), by = 0.1),
  p = burst_effect(ic80)
) %>%
  ggplot(aes(x = ic80, y = p)) +
  geom_line() +
  scale_x_log10() +
  labs(y = "Avg. infectious virions / cell day")

```

```{r placebo-fp-effect}

fp_placebo_data = vl_sims_placebo %>%
  filter(first_positive_row)

fp_placebo_data_wide = fp_placebo_data %>%
  select(pub_id, indirect, log10V, gmt80ls, initT_mode) %>%
  pivot_wider(values_from = log10V, names_from = indirect, names_prefix = "indirect")

fp_bp_placebo = fp_placebo_data %>%
  mutate(ic80_cat2 = factor(ic80_cat, 
                            levels = c("IC80 < 1 mcg/mL", "IC80 >= 1 mcg/mL"),
                            labels = c("VRC01 sensitive: IC80<1", 
                                       "VRC01 resistant: IC80\u22651"))
         ) %>%
  ggplot(aes(x = factor(indirect), y = log10V)) +
  geom_boxplot(width = 0.2,) +
  geom_point(aes(color = ic80_cat), position = position_jitter(width = 0.1), alpha = 0.4) +
  geom_path(aes(group = pub_id, color = ic80_cat), alpha = 0.25) +
  scale_color_manual(values = ic80gte1_colors) +
  ggplot2::facet_wrap(vars(ic80_cat2)) +
  scale_x_discrete(limits = c(F, T), labels = c("Median IC80", "Obs. IC80")) +
  labs(x = "Indirect effect on burst size", 
       y = expression(paste("Predicted first positive viral load log"[10], "(copies/mL)"))) +
  theme(strip.background = element_blank(), panel.grid = element_blank(),
        strip.text.x = element_text(face = "bold"), legend.position = "none") 

# cairo_pdf("06_amp-sim-analysis_files/figure-html/fp-placebo-comparison-1.pdf", family="DejaVu Sans")
# fp_bp_placebo + theme(aspect.ratio = 3/4)
# dev.off()

fp_bp_placebo

```

```{r placebo-indirect}

ic80_range_burst_size = c(round(placebo_ic80_range, 2), 1, 10)

burst_size_models = map_df(ic80_range_burst_size, function(i){
  theta_vl = theta_vl_placebo
  theta_vl$p = burst_effect(i)
  
  run_pkpd_models(
        0:100,
        holte_pkpd_model,
        theta_pk = theta_pk,
        theta_vl = theta_vl,
        theta_pd = tibble(IC80 = 4, rho = Inf, h = 1),
        infection_time = 0,
        initial_pk_dose = 0
      ) %>% 
        mutate(isolate_ic80 = i, burst_size = theta_vl$p)
  })

placebo_adj_vl_pl = ggplot(burst_size_models, aes(x = time/7, y = log10V, 
                                                  colour = factor(isolate_ic80))) +
  geom_line() +
  labs(x = "Weeks post-acquisition", y = "Placebo-population log10 VL",
       colour = "VRC01 IC80 of acquired isolate") +
  scale_color_viridis_d() +
  scale_y_continuous(breaks = -1:8)

placebo_adj_vl_pl
# inset the ic80 vs pi plot # 2 x 2
# 4 x 3

```

# VRC01 analysis

## Varying VRC01 effects

```{r indiv-sim-curves}

sim_plot = vl_sims_vrc01 %>%
  ggplot(aes(x = days_fp, y = pmax(1, log10V), colour = ic80_cat)) +
  geom_line(aes(group = pub_id), alpha = 0.25) +
  geom_boxplot(data = filter(vl_sims_vrc01, first_positive_row), 
               width = 10, outlier.colour = NULL, fill = NA, colour = "black",
               aes(x = 0)) +
  scale_color_manual(values = ic80gte1_colors) +
  scale_x_continuous(breaks = 20 * (-1:3)) +
  scale_y_continuous(breaks = 2 * 1:4, limits = c(NA, 8)) +
  lemon::facet_rep_grid(rows = vars(ic80_cat2), cols = vars(sim_type2),
                        repeat.tick.labels = FALSE) +
  labs(x = "Time (days relative to first positive)", 
       y = expression(paste("Simulated viral load log"[10], "(copies/mL)"))) +
  coord_cartesian(xlim = c(-20, 60)) +
  theme(legend.position = "none", strip.background = element_blank(), panel.grid = element_blank(),
        panel.spacing = unit(0, "lines"),
        text = element_text(family = "sans"), 
        strip.text.y = element_text(face = "bold", size = 6), 
        strip.text.x = element_text(face = "bold", size = 8), 
        strip.clip = "off"
         )


cairo_pdf("06_amp-sim-analysis_files/figure-html/indiv-sim-curves-1.pdf", family="DejaVu Sans")
sim_plot + theme(aspect.ratio = 0.9)
dev.off()

sim_plot
  
```
 
```{r fp-comparison}

fp_bp = vl_sims_vrc01 %>%
  filter(first_positive_row) %>%
  ggplot(aes(x = sim_type2, y = log10V)) +
  geom_boxplot(width = 0.2,) +
  geom_point(aes(color = ic80_cat), position = position_jitter(width = 0.1), alpha = 0.4) +
  geom_path(aes(group = pub_id, color = ic80_cat), alpha = 0.25) +
  scale_color_manual(values = ic80gte1_colors) +
  ggplot2::facet_wrap(vars(ic80_cat2)) +
  labs(x = "", 
       y = expression(paste("Predicted first positive viral load log"[10], "(copies/mL)"))) +
  theme(strip.background = element_blank(), panel.grid = element_blank(),
        strip.text.x = element_text(face = "bold"), legend.position = "none") 

# cairo_pdf("06_amp-sim-analysis_files/figure-html/fp-comparison-1.pdf", family="DejaVu Sans")
# fp_bp + theme(aspect.ratio = 3/4)
# dev.off()

fp_bp

```

```{r fp-test-by-design}
paper_comparisons = c("VRC01 (AMP design) vs. No VRC01",
                "VRC01 (AMP design) vs. 10x potency",
                "No VRC01 vs. 10x potency",
                "10x potency vs. 100x potency",
                "100x potency vs. 1000x potency")

sim_fp_tests = vl_sims_vrc01 %>%
  filter(first_positive_row) %>%
  group_by(ic80_cat) %>%
  summarize(VISCfunctions::pairwise_test_cont(
    x = log10V,
    group = sim_type2,
    paired = T,
    id = pub_id,
    digits = 2
  ),
      .groups = "keep"
) %>%
  mutate(
    significant = MagnitudeTest < 0.05
  ) %>%
  select(-Mean_SD, -PerfectSeparation) %>%
  rename(signed_rank_pvalue = MagnitudeTest, N = SampleSizes) %>%
  filter(Comparison %in% paper_comparisons)


sim_fp_tests %>%
  write_csv("06_amp-sim-analysis_files/figure-html/sim-fp-comparisons.csv") %>%
  kable(caption = "Tests of first positives across potency") %>%
  kable_styling(full_width = F)

```

```{r fp-tests-sensitivity}

sim_fp_sens_tests = vl_sims_vrc01 %>%
  filter(first_positive_row) %>%
  group_by(sim_type2) %>%
  summarize(VISCfunctions::pairwise_test_cont(
    x = log10V,
    group = ic80_cat,
    paired = F,
    digits = 2
  ),
      .groups = "keep"
) %>%
  mutate(
    significant = MagnitudeTest < 0.05
  ) %>%
  select(-Mean_SD, -PerfectSeparation) %>%
  rename(wilcox_pvalue = MagnitudeTest, N = SampleSizes)


sim_fp_sens_tests %>%
  kable(caption = "Tests of first positives within potency between sensitivity") %>%
  kable_styling(full_width = F)

```

## Dose-response

Seems very minimal in the range of the data

```{r dr-setup}
id80_max = conc_range[2]/placebo_ic80_range[1]
id50_max = conc_range[2]/placebo_ic50_range[1]

id50_max_vrc01 = conc_range[2]/vrc01_ic50_range[1]
id80_max_vrc01 = 4^(-1/amp_theta_pd$h) * id50_max_vrc01

dose_response = tibble(
  #id80 = seq(id80_range[1], id80_range[2], by = 0.1),
  id50 = 10^seq(-1, log10(id50_max), by = 0.1),
  PT80 = 4^(-1/amp_theta_pd$h) * id50,
  in_vitro = 1/(1+(id50)^(-amp_theta_pd$h)),
  in_vivo = 1/(1+(id50/amp_theta_pd$rho)^(-amp_theta_pd$h)),
  pe = predict(empirical_amp_fit, data.frame(PT80 = PT80))
)
  
```

```{r dr-plt}
pkpd_dr_pl = dose_response %>%
  gather("type", "neutralization", in_vitro, in_vivo) %>%
  mutate(iip = -log10(1 - neutralization)) %>%
  gather("outcome", "value", neutralization, iip, factor_key = T) %>%
  ggplot(aes(x = PT80, y = value)) +
  geom_line(aes(colour = type)) +
  scale_x_log10() +
  annotate("rect", xmin = id80_max_vrc01, xmax = Inf, ymin = -Inf, ymax = Inf, alpha =0.5, 
           fill = "gray") +
  annotate("text", label = 'placebo\nIC80\nonly', x = 7000, y = 0.65) +
  labs(y = "", colour = "VRC01 PKPD Activity", x = "Theoretical AMP PT80 Range") +
  geom_vline(data = pt80_efficacy, aes(xintercept = PT80), linetype = "dashed") +
  geom_text(data = pt80_efficacy, aes(label = PE_pct), angle = 90, y = 0.65, vjust = 0) +
  facet_wrap(~outcome, scales = "free_y", nrow = 1, strip.position = "left") +
  strip_as_yaxis

pkpd_dr_pl

```


```{r, eval = F}

pkpd_dr_pl_alt = dose_response %>%
  gather("type", "neutralization", in_vitro, in_vivo, pe) %>%
  mutate(
    iip = -log10(1 - neutralization),
    percentage = neutralization * 100
    ) %>%
  gather("outcome", "value", percentage, iip, factor_key = T) %>%
  mutate(
    activity = factor(type, levels = c("in_vitro", "in_vivo", "pe"),
                      labels = c("Neutralization (in vitro)", "PKPD (in vivo)", "PE"))
  ) %>%
  ggplot(aes(x = PT80, y = value)) +
  geom_line(aes(colour = activity)) +
  scale_x_log10() +
  annotate("rect", xmin = id80_max_vrc01, xmax = Inf, ymin = -Inf, ymax = Inf, alpha =0.5, 
           fill = "gray") +
  annotate("text", label = 'placebo\nIC80\nonly', x = 7000, y = 5) +
  labs(y = "", colour = "VRC01 Activity", x = "Theoretical AMP PT80 Range") +
  facet_wrap(~outcome, scales = "free_y", nrow = 1, strip.position = "left") +
  strip_as_yaxis

pkpd_dr_pl_alt

```

## Population PKPD Effect on Viral Load

```{r setup}

id80_range = c(0, 10, 200, 1000, 2500, 10000)

```

In the context of AMP, need to show the effect of titer on viral load,

- show how R0 and titer are related at time of infection
- within AMP pick different levels of titer, using VRC01 PK with and without burst size adjustment

### R0 vs. titer

This does not adjust for burst size. If this wer eincluded, at a fixed titer, lower IC80 would have lower extinction probability (higher R0).

```{r r0-titer}

# 0.329877
# theta_vl_vrc01$Bt0 * 1/(1 + (1000/(0.329877 * amp_theta_pd$rho))^amp_theta_pd$h)
# 4.995823e-06 * theta_vl_vrc01$p * theta_vl_vrc01$aS / (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23)


r0_titer_sim = tibble(
  PT80 = 1:10000,
  ID50 = PT80 * 4 ^(1/amp_theta_pd$h),
  amp_rho = amp_theta_pd$rho,
  pegu_rho = 100
  ) %>%
  gather("potency_reduction", "rho", amp_rho, pegu_rho) %>%
  mutate(
    beta_neut = beta_titer_neut(ID50/rho, amp_theta_pd$h),
    R0 = beta_neut*theta_vl_vrc01$Bt0 * theta_vl_vrc01$p * theta_vl_vrc01$aS /
        (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23),
    extinction_prob = pmin(1, 1/R0)
    ) 

r0_titer_sim %>%
  ggplot(aes(x = PT80, y = R0)) +
  geom_line(aes(colour = potency_reduction)) +  
  geom_hline(yintercept = 1, linetype = "dashed") +
  scale_x_log10()


r0_titer_sim %>%
  ggplot(aes(x = PT80, y = extinction_prob)) +
  geom_line(aes(colour = potency_reduction)) +  
  geom_vline(data = pt80_efficacy, aes(xintercept = PT80), linetype = "dashed") +
  geom_text(data = pt80_efficacy, aes(label = PE_pct), angle = 90, y = 0.65, vjust = 0) +
  scale_x_log10()


```

### PKPD Trajectories with fixed PT80

Ie, if there were constant levels of bnab with fixed PT80. (no indirect effect)

```{r amp-pkpd-pt80-trajectories}

amp_titer_models = map_df(id80_range, function(pt80){
  amp_theta_pd$IC80 = 1
  initial_dose = pt80 * theta_pk$V1
  theta_pk$Cl = 0
  theta_pk$Q = 0
  
  run_pkpd_models(
    c(seq(0, 1, by = 0.1), 2:100),
    holte_pkpd_model,
    theta_pk = theta_pk,
    theta_vl = theta_vl_vrc01,
    theta_pd = amp_theta_pd,
    infection_time = 0,
    initial_pk_dose = initial_dose
  ) %>%
    mutate(
      IC80 = 1,
      PT80 = pt80,
      min_VL = min(log10V)
    )
})

ggplot(amp_titer_models, aes(x = time/7, y = pmax(0.01, centr/IC80), colour = factor(PT80))) +
  scale_y_log10("PT80", breaks = c(0.01, 10, 100, 1000), 
                labels = c(0, 10, 100, 1000)) +
  geom_line() +
  labs( x = "Weeks relative to crossing detection threshold", colour = "PT80 at infection")

amp_titer_models %>%
  ggplot(aes(x = time/7, y = pmax(log10(20), log10V), 
                             colour = factor(PT80))) +
  scale_linetype_discrete(breaks = T, labels = "Potential suppression") +
  scale_y_continuous(breaks = c(log10(20), 2:7), labels = c("< LoD", 2:7)) +
  labs(y = "Log10 VL", x = "Weeks", colour = "Constant PT80", 
       linetype = "") +
  geom_line() 

```

### PKPD Trajectories with PK

Direct effect only. PK follows VRC01. Does not currently show AMP dose but a PT80 of 1000 would be like 
[1000 ug/mL] / IC80 of 1, roughly the peak of high dose

```{r amp-pkpd-trajectories}

amp_pkpd_models = map_df(id80_range, function(pt80){
  amp_theta_pd$IC80 = 1
  initial_dose = pt80 * theta_pk$V1
  
  run_pkpd_models(
    c(seq(0, 1, by = 0.1), 2:100),
    holte_pkpd_model,
    theta_pk = theta_pk,
    theta_vl = theta_vl_vrc01,
    theta_pd = amp_theta_pd,
    infection_time = 0,
    initial_pk_dose = initial_dose
  ) %>%
    mutate(
      IC80 = 1,
      PT80 = pt80,
      min_VL = min(log10V),
      min_detection_time = min(time[log10V > log10(20)]),
      R0 = theta_vl_vrc01$Bt0 * theta_vl_vrc01$p * theta_vl_vrc01$aS /
        (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23),
      R0_adj = Bt0_adj[1] * theta_vl_vrc01$p * theta_vl_vrc01$aS /
        (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23)
    )
})

ggplot(amp_pkpd_models, aes(x = time, y = centr/IC80, colour = factor(PT80))) +
  scale_y_log10("PT80", breaks = c(0.1, 10, 1000)) +
  geom_line()

# ggplot(amp_pkpd_models, aes(x = (time - min_detection_time)/7, y = centr/IC80, colour = factor(PT80))) +
#   scale_y_log10("PT80", breaks = c(0.1, 10, 1000)) +
#   geom_line() +
#   labs( x = "Weeks relative to crossing detection threshold", colour = "PT80 at infection")

amp_pkpd_models %>%
  mutate(extinction = min_VL < -2) %>%
  ggplot(aes(x = time/7, y = pmax(-2, log10V), 
                             colour = factor(PT80),
                             linetype = factor(extinction))) +
  scale_linetype_discrete(breaks = T, labels = "Potential suppression") +
  labs(y = "Log10 VL", x = "Weeks", colour = "PT80 at infection", 
       linetype = "") +
  geom_line() 
# 
# 
# amp_pkpd_models %>%
#   mutate(extinction = min_VL < -2) %>%
#   ggplot(aes(x = (time - min_detection_time)/7, y = pmax(-2, log10V), 
#                              colour = factor(PT80),
#                              linetype = factor(extinction))) +
#   scale_linetype_discrete(breaks = T, labels = "Potential suppression") +
#   scale_y_continuous(breaks = -1:7) +
#   labs(y = "Log10 VL", x = "Weeks relative to crossing detection threshold", colour = "PT80 at infection", 
#        linetype = "") +
#   geom_line() 

```

### PKPD Trajectories with indirect effects with PK

The PK of the ID80 titers are fixed, but they are calculated using different input IC80 because that affects burst size here.

```{r pkpd-indirect}

ic80_range = c(0.1, 1, 10)

amp_pkpd_indirect_models = map_df(id80_range, function(pt80){
  map_df(ic80_range, function(ic80){
    
    amp_theta_pd$IC80 = ic80
    theta_vl_vrc01$p = burst_effect(ic80)
    
    initial_dose = pt80 * ic80 * theta_pk$V1
  
    run_pkpd_models(
      c(seq(0, 1, by = 0.1), 2:100),
      holte_pkpd_model,
      theta_pk = theta_pk,
      theta_vl = theta_vl_vrc01,
      theta_pd = amp_theta_pd,
      infection_time = 0,
      initial_pk_dose = initial_dose
    ) %>%
      mutate(
        IC80 = ic80,
        IC80_lab = paste("IC80 =", ic80),
        PT80 = pt80,
        min_VL = min(log10V),
        min_detection_time = min(time[log10V > log10(20)]),
        R0 = theta_vl_vrc01$Bt0 * theta_vl_vrc01$p * theta_vl_vrc01$aS /
          (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23),
        R0_adj = Bt0_adj[1] * theta_vl_vrc01$p * theta_vl_vrc01$aS /
          (theta_vl_vrc01$dS * theta_vl_vrc01$dI * 23)
      )
  })
})

ggplot(amp_pkpd_indirect_models, aes(x = time, y = centr/IC80, colour = factor(PT80))) +
  scale_y_log10("PT80", breaks = c(0.1, 10, 1000)) +
  facet_wrap(~IC80_lab)+
  geom_line()

amp_pkpd_indirect_models %>%
  mutate(extinction = min_VL < -2) %>%
  ggplot(aes(x = time/7, y = pmax(-2, log10V), 
                             colour = factor(PT80),
                             linetype = factor(extinction))) +
  scale_linetype_discrete(breaks = T, labels = "Potential suppression") +
  labs(y = "Log10 VL", x = "Weeks", colour = "PT80 at infection", 
       linetype = "") +
    facet_wrap(~IC80_lab)+
  geom_line() 

amp_pkpd_indirect_models %>%
  mutate(
    extinction = min_VL < -2,
    PT80_lab = paste("Infection PT80 =", PT80)) %>%
  ggplot(aes(x = time/7, y = pmax(-2, log10V), 
                             colour = factor(IC80),
                             linetype = factor(extinction))) +
  scale_linetype_discrete(breaks = T, labels = "Potential suppression") +
  labs(y = "Log10 VL", x = "Weeks", colour = "IC80 at infection", 
       linetype = "") +
    facet_wrap(~PT80_lab)+
  geom_line() 

```

### Simulated PKPD AMP Summary Measures

Attempt to recreate first positive.

```{r amp-pkpd-sim-summary}

calc_delay = 1
auc_end = 90
fine_id80_range = 10^seq(-0.25, log10(id80_max_vrc01), by = 0.1)

amp_titer_vl_summary = map_df(fine_id80_range, function(pt80){
  amp_theta_pd$IC80 = 1
  initial_dose = pt80 * theta_pk$V1
  
  run_pkpd_models(
    c(seq(0, 20, by = 0.1), 21:150),
    holte_pkpd_model,
    theta_pk = theta_pk,
    theta_vl = theta_vl_vrc01,
    theta_pd = amp_theta_pd,
    infection_time = 0,
    initial_pk_dose = initial_dose
  ) %>%
    mutate(
      IC80 = 1,
      PT80 = pt80,
      min_VL = min(log10V),
      min_detection_time = min(time[log10V > log10(20)]),
      PT80_detection = centr[min_detection_time],
      PT80_2wks = 10^mean(log10(centr[time >= min_detection_time & time < (min_detection_time + 14)]))
    )
}) %>%
  group_by(PT80, PT80_detection, PT80_2wks, min_detection_time) %>%
  summarize(
    avg_2wk = mean(log10V[time >= calc_delay & time <= 14+calc_delay]),
    geo_auc_3mo = tail(pkr::AUC(time[time >= calc_delay & time <= auc_end + calc_delay], 
                                log10V[time >= calc_delay & time <= auc_end + calc_delay])[,1], 1)/auc_end,
    avg_2wk_adj = mean(log10V[time >= min_detection_time & time < (min_detection_time + 14)]),
    geo_auc_3mo_adj = tail(pkr::AUC(time[time >= min_detection_time &
                                       time <= auc_end + min_detection_time], 
                                log10V[time >= min_detection_time &
                                       time <= auc_end + min_detection_time])[,1], 1)/auc_end,
    .groups = "drop"
  )

amp_titer_vl_summary %>%
  gather("measure", "value", avg_2wk_adj) %>%
  ggplot(aes(x = PT80_2wks, y = value)) +
  geom_line() +
  geom_point(data = vrc01_first_pos, aes(x = pmax(0.1, PT80), y = log10vl),
             alpha = 0.5) +
scale_x_log10() +
  labs(y = "predicted first positive", x = "predicted PT80 at first positive") 

```

# Save data

```{r, eval = !rerun_sims}
check_first_pos = vl_sims_vrc01 %>%
  filter(first_positive_row) %>%
  group_by(sim_type) %>%
  summarize(total = n_distinct(pub_id), .groups = "drop")
stopifnot(all(check_first_pos$total == 96))

sims_out = vl_sims_vrc01 %>%
  select(pub_id, sim_type, days_fp, log10V, ic80_cat, contains("gmt"), first_positive_row, fp_day, 
         model_infection_day, days_enrollment)

write_csv(sims_out, clean_data_here("amp-individual-vrc01-sims.csv"))

```

```{r placebo-save, eval = save_placebo}
check_first_pos = vl_sims_placebo %>%
  filter(first_positive_row) %>%
  group_by(indirect) %>%
  summarize(total = n_distinct(pub_id), .groups = "drop")
stopifnot(all(check_first_pos$total == 62))

sims_out_placebo = vl_sims_placebo %>%
  filter(indirect) %>%
  select(pub_id, rx_code2, indirect, days_fp, log10V, ic80_cat, contains("gmt"), first_positive_row, fp_day, 
         model_infection_day, days_enrollment)

write_csv(sims_out_placebo, clean_data_here("amp-individual-placebo-sims.csv"))

```


```{r save-shared, eval = save_shared_folder}

write_csv(tribble(
  ~variable_name, ~description,
  "sim_type", "NoTreatment = VRC01 off, VRC01 was AMP, VRC01xX has X is multiplier",
  "days_fp", "days relative to first positive",
  "ic80_cat", "binary category based on gmt80ls",
  "first_positive_row", "subset on this to get the observation time of the first positive"),
  paste0("/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/!INDIVIDUAL SIMS - ",
         Sys.Date()))

write_csv(sims_out, "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/amp-individual-vrc01-sims.csv")

```

```{r save-shared-placebo, eval = save_shared_folder_placebo}

write_csv(tribble(
  ~variable_name, ~description,
  "days_fp", "days relative to first positive",
  "ic80_cat", "binary category based on gmt80ls",
  "first_positive_row", "subset on this to get the observation time of the first positive"),
  paste0("/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/!PLACEBO INDIVIDUAL SIMS - ",
         Sys.Date()))

write_csv(sims_out_placebo, "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/amp-individual-placebo-sims.csv")

```
